Remove version numbers
The project is currently too unstable to use semantic versioning. Attempting semver is
preventing rapid development, so for the moment it is being entirely removed.

Versioning may be re-added at a later stage in the project when the codebase is more
mature and stable, but for the moment it's getting in the way of putting things out there
quickly.